`dboy r` robustnessness on restarting images.
I believe it's recommended to start the original container with a living command such as bash or python.
I believe if you do not do this, when you try to start an already existing container, it will die instantly.
If this is the case, this current functionality is a hack. The proper solution would be committing the stopped
container and rebuilding an image from it under the same name, or to simply delete and rebuild from the base.